backend/requirements.txt
# FastAPI Framework
fastapi>=0.103.1
pydantic>=2.4.1
pydantic-settings>=2.0.3
email-validator>=2.0.0

# ASGI Server
uvicorn>=0.23.2
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.21
alembic>=1.12.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.7

# Authentication
python-jose>=3.3.0
pyjwt>=2.8.0
cryptography>=41.0.3

# AWS
boto3>=1.28.48
botocore>=1.31.48

# File handling
python-multipart>=0.0.6
qrcode>=7.4.2
pillow>=10.0.1

# HTTP Client
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic[email]>=2.4.1
ujson>=5.8.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1